runtimes/agent-base/pyproject.toml
[project]
name = "agent-base"
version = "0.1.0"
description = "Base runtime for Agent-Pools — Lambda-style dynamic code loader (LangGraph / ADK / custom)"
requires-python = ">=3.12,<3.13"
dependencies = [
    "runtime-common",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Frameworks the base image is expected to host. Pin minimum versions here;
    # individual deployed bundles may still vendor their own pins via the bundle loader.
    "langgraph>=1.0",
    "deepagents>=0.5",
    # init_chat_model auto-dispatches by model spec prefix; provider deps must
    # be installed for whichever the bundle picks. Add new langchain-* here as
    # we support more providers.
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "google-adk>=0.1.0 ; python_version >= '3.12'",
    "opik>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_base"]
